% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testc-pkg.R
\docType{package}
\name{testc-package}
\alias{testc}
\alias{testc-package}
\title{testc: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
